# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClinicFlow Hospital & Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments with optional filters
      parameters:
        - in: query
          name: doctorId
          schema:
            type: integer
        - in: query
          name: patientId
          schema:
            type: integer
        - in: query
          name: date
          schema:
            type: string
            format: date
        - in: query
          name: status
          schema:
            type: string
            enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/today:
    get:
      operationId: getTodayAppointments
      tags: [appointments]
      summary: Get today's appointments grouped by status
      parameters:
        - in: query
          name: doctorId
          schema:
            type: integer
      responses:
        "200":
          description: Today's appointments summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayAppointmentsSummary"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment (status, time, etc.)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel/delete appointment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── QUEUE ────────────────────────────────────────────────────────────────────
...
  /consultations:
    get:
      operationId: listConsultations
      tags: [consultations]
      summary: List consultations
      parameters:
        - in: query
          name: patientId
          schema:
            type: integer
        - in: query
          name: doctorId
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of consultations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultation"
    post:
      operationId: createConsultation
      tags: [consultations]
      summary: Start a consultation for an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConsultationBody"
      responses:
        "201":
          description: Consultation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /consultations/{id}:
    get:
      operationId: getConsultation
      tags: [consultations]
      summary: Get consultation by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
        "404":
          description: Not found
    patch:
      operationId: updateConsultation
      tags: [consultations]
      summary: Update consultation notes/diagnosis
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConsultationBody"
      responses:
        "200":
          description: Updated consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"

  # ─── PRESCRIPTIONS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
        type:
          type: string
          enum: [routine, follow_up, emergency, consultation]
        notes:
          type: ["string", "null"]
        queueNumber:
          type: ["integer", "null"]
        patientName:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, patientId, doctorId, scheduledAt, status, type, createdAt, updatedAt]
    CreateAppointmentBody:
      type: object
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
        type:
          type: string
          enum: [routine, follow_up, emergency, consultation]
        notes:
          type: ["string", "null"]
      required: [patientId, doctorId, scheduledAt, type]
    UpdateAppointmentBody:
      type: object
      properties:
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
        type:
          type: string
          enum: [routine, follow_up, emergency, consultation]
        notes:
          type: ["string", "null"]
        queueNumber:
          type: ["integer", "null"]
    TodayAppointmentsSummary:
      type: object
      properties:
        date:
          type: string
        total:
          type: integer
        scheduled:
          type: integer
        confirmed:
          type: integer
        inProgress:
          type: integer
        completed:
          type: integer
        cancelled:
          type: integer
        noShow:
          type: integer
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
      required: [date, total, scheduled, confirmed, inProgress, completed, cancelled, noShow, appointments]

    # ─── QUEUE ────────────────────────────────────────────────────────────────
...
    Consultation:
      type: object
      properties:
        id:
          type: integer
        appointmentId:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        chiefComplaint:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        treatmentPlan:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        vitals:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [in_progress, completed]
        patientName:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, appointmentId, patientId, doctorId, status, createdAt, updatedAt]
    CreateConsultationBody:
      type: object
      properties:
        appointmentId:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        chiefComplaint:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        treatmentPlan:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        vitals:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
      required: [appointmentId, patientId, doctorId]
    UpdateConsultationBody:
      type: object
      properties:
        chiefComplaint:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        treatmentPlan:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        vitals:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [in_progress, completed]

    # ─── PRESCRIPTION ─────────────────────────────────────────────────────────
...